% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextractr-class.R
\name{Contextractr$locate_keywords}
\alias{Contextractr$locate_keywords}
\title{Locate/classify columns according to keywords}
\arguments{
\item{.df}{the dataframe to process}

\item{selection}{the unquoted column for contextractr to look at}
}
\value{
the input dataframe, with extra columns appended corresponding to
the keywords found
}
\description{
Locate/classify columns according to keywords
}
\seealso{
Other Contextractr: \code{\link{Contextractr$add_json}},
  \code{\link{Contextractr$add_serial}},
  \code{\link{Contextractr$new}},
  \code{\link{Contextractr}}
}
\concept{Contextractr}
