% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextractr-class.R
\name{Contextractr$new}
\alias{Contextractr$new}
\title{Create a new instance of Contextractr}
\description{
Provide a mapping path, mapping list, or mapping tibble to allow setting of
the mapping dictionary
}
\seealso{
Other Contextractr: \code{\link{Contextractr$add_json}},
  \code{\link{Contextractr$add_serial}},
  \code{\link{Contextractr$add_yaml}}
}
\concept{Contextractr}
