% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextractr-class.R
\name{Contextractr}
\alias{Contextractr}
\title{Contextractr - Fuzzy Extraction and Classification by Context}
\description{
This R6 class aims to provide a simple interface for providing
pattern + context based rules in order to extract values by
keywords
}
\seealso{
Other Contextractr: \code{\link{Contextractr$add_json}},
  \code{\link{Contextractr$add_serial}},
  \code{\link{Contextractr$locate_keywords}},
  \code{\link{Contextractr$new}}
}
\concept{Contextractr}
\concept{Contextractr-important}
