% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{csv_to_serial}
\alias{csv_to_serial}
\title{Convert a CSV into a serial mapping suitable for Contextractr}
\usage{
csv_to_serial(df)
}
\arguments{
\item{df}{the input dataframe, more or less a raw csv}
}
\value{
the input df, serialized.
}
\description{
Given a csv, with one row per keyword + approx.match specifier, grouped
according to "title" (Single value for title may span multiple keywords)
transform this to something digestable by contextractr.
}
\examples{
if (requireNamespace("readr", quietly = TRUE)){
  in_csv <- readr::read_csv("../../tests/testthat/csv_specifier.csv")
  print(in_csv)

  out <- in_csv \%>\% csv_to_serial()
  if (requireNamespace("utils", quietly = TRUE)) utils::str(out)
}
}
\concept{IO}
