% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextractr-class.R
\name{Contextractr$add_json}
\alias{Contextractr$add_json}
\title{Add a json mapping}
\description{
Add a json mapping
}
\seealso{
Other Contextractr: \code{\link{Contextractr$add_serial}},
  \code{\link{Contextractr$add_yaml}},
  \code{\link{Contextractr$new}}
}
\concept{Contextractr}
