% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextractr-class.R
\name{Contextractr$add_serial}
\alias{Contextractr$add_serial}
\title{Add a serial mapping}
\description{
Add a serial mapping
}
\seealso{
Other Contextractr: \code{\link{Contextractr$add_json}},
  \code{\link{Contextractr$add_yaml}},
  \code{\link{Contextractr$new}}
}
\concept{Contextractr}
