% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextractr-class.R
\name{Contextractr$add_yaml}
\alias{Contextractr$add_yaml}
\title{Add a yaml mapping}
\description{
Add a yaml mapping
}
\seealso{
Other Contextractr: \code{\link{Contextractr$add_json}},
  \code{\link{Contextractr$add_serial}},
  \code{\link{Contextractr$new}}
}
\concept{Contextractr}
